定时任务

+++ src/main/java/io/cem/modules/cem/entity/RecordDayFtpEntity.java
@@ -32,7 +32,7 @@ public class RecordDayFtpEntity implements Serializable {
 	//测试目标ID
 	private Integer targetId;
 	//测试目标IP
-	private Integer targetIp;
+	private Long targetIp;
 	//记录日期
 	private Date recordDate;
 	//记录时间
@@ -151,13 +151,13 @@ public class RecordDayFtpEntity implements Serializable {
 	/**
 	 * 设置：测试目标IP
 	 */
-	public void setTargetIp(Integer targetIp) {
+	public void setTargetIp(Long targetIp) {
 		this.targetIp = targetIp;
 	}
 	/**
 	 * 获取：测试目标IP
 	 */
-	public Integer getTargetIp() {
+	public Long getTargetIp() {
 		return targetIp;
 	}
 	/**
